add basic moveit config
