+++ README.md
@@ -1,9 +1,22 @@
 # ovis-ros2
 
-Launch gz sim :
+Ovis is a 6 Degrees of Freedom (DoF) robotic arm developed by the CAPRA student club. It is designed to be highly versatile and is compatible with MoveIt for motion planning and Kinova motors for precise control. This repository provides the necessary tools and launch files to simulate and visualize the Ovis robotic arm using ROS2.
+
+
+## Installation
+
+Clone and build the repository :
 ```bash
+git clone https://github.com/clubcapra/ovis_ros2.git
+cd ovis_ros2
 colcon build --symlink-install
 source install/setup.bash
+```
+
+## Usage
+
+Start the gazebo simulation :
+```bash
 ros2 launch ovis_bringup sim.launch.py
 ```
 
@@ -12,4 +25,9 @@ Launch gui state publisher and arm visualization :
 colcon build --symlink-install
 source install/setup.bash
 ros2 launch ovis_description launch.py
+```
+
+Launch MoveIt! :
+```bash
+ros2 launch ovis_moveit demo.launch.py
 ```
