+++ src/ovis_description/urdf/gazebo.urdf.xacro
@@ -5,7 +5,7 @@
 
         <!-- ros_control plugin -->
         <gazebo>
-            <plugin name="gazebo_ros_control" filename="libgazebo_ros_control.so">
+            <plugin name="gazebo_ros_control" filename="gz_ros_control.so">
                 <robotNamespace>${robot_namespace}</robotNamespace>
                 <robotSimType>gazebo_ros_control/DefaultRobotHWSim</robotSimType>
                 <legacyModeNS>true</legacyModeNS>
