+++ src/ovis_description/urdf/kinova_common.xacro
@@ -112,7 +112,7 @@
         <provideFeedback>true</provideFeedback>
       </gazebo>
       <gazebo>
-        <plugin name="ft_sensor" filename="libgazebo_ros_ft_sensor.so">
+        <plugin name="ft_sensor" filename="gz_ros_ft_sensor.so">
           <updateRate>30.0</updateRate>
           <topicName>${joint_name}_ft_sensor_topic</topicName>
           <jointName>${joint_name}</jointName>
