+++ src/ovis_description/urdf/ovis_standalone.urdf.xacro
@@ -1,34 +1,18 @@
 <?xml version="1.0"?>
 
-<robot xmlns:xi="http://www.w3.org/2001/XInclude"
-    xmlns:gazebo="http://playerstage.sourceforge.net/gazebo/xmlschema/#gz"
-    xmlns:model="http://playerstage.sourceforge.net/gazebo/xmlschema/#model"
-    xmlns:sensor="http://playerstage.sourceforge.net/gazebo/xmlschema/#sensor"
-    xmlns:body="http://playerstage.sourceforge.net/gazebo/xmlschema/#body"
-    xmlns:geom="http://playerstage.sourceforge.net/gazebo/xmlschema/#geom"
-    xmlns:joint="http://playerstage.sourceforge.net/gazebo/xmlschema/#joint"
-    xmlns:controller="http://playerstage.sourceforge.net/gazebo/xmlschema/#controller"
-    xmlns:interface="http://playerstage.sourceforge.net/gazebo/xmlschema/#interface"
-    xmlns:rendering="http://playerstage.sourceforge.net/gazebo/xmlschema/#rendering"
-    xmlns:renderable="http://playerstage.sourceforge.net/gazebo/xmlschema/#renderable"
-    xmlns:physics="http://playerstage.sourceforge.net/gazebo/xmlschema/#physics"
-    xmlns:xacro="http://www.ros.org/wiki/xacro" name="ovis">
+<robot name="ovis" xmlns:xacro="http://www.ros.org/wiki/xacro">
 
 
+
+    <xacro:include filename="$(find ovis_description)/urdf/gazebo.urdf.xacro" />
     <xacro:include filename="$(find ovis_description)/urdf/ovis.urdf.xacro" />
+    <xacro:include filename="$(find ovis_description)/urdf/kinova_common.xacro" />
 
+    
     <link name="base_link" />
 
-    <!-- for gazebo -->
-    <!-- <link name="world" /> -->
-
-    <!-- <joint name="connect_root_and_world" type="fixed">
-        <child link="root" />
-        <parent link="world" />
-        <origin xyz="0 0 0" rpy="0 0 0" />
-    </joint> -->
-
     <xacro:property name="robot_root" value="base_link" />
+
 
     <xacro:ovis base_parent="${robot_root}" />
