+++ src/ovis_moveit/config/initial_positions.yaml
@@ -1,0 +1,9 @@
+# Default initial positions for ovis's ros2_control fake system
+
+initial_positions:
+  ovis_joint_1: 0
+  ovis_joint_2: 0
+  ovis_joint_3: 0
+  ovis_joint_4: 0
+  ovis_joint_5: 0
+  ovis_joint_6: 0
