+++ src/ovis_moveit/config/joint_limits.yaml
@@ -1,0 +1,40 @@
+# joint_limits.yaml allows the dynamics properties specified in the URDF to be overwritten or augmented as needed
+
+# For beginners, we downscale velocity and acceleration limits.
+# You can always specify higher scaling factors (<= 1.0) in your motion requests.  # Increase the values below to 1.0 to always move at maximum speed.
+default_velocity_scaling_factor: 0.1
+default_acceleration_scaling_factor: 0.1
+
+# Specific joint properties can be changed with the keys [max_position, min_position, max_velocity, max_acceleration]
+# Joint limits can be turned off with [has_velocity_limits, has_acceleration_limits]
+joint_limits:
+  ovis_joint_1:
+    has_velocity_limits: true
+    max_velocity: 0.62831853071795862
+    has_acceleration_limits: false
+    max_acceleration: 0
+  ovis_joint_2:
+    has_velocity_limits: true
+    max_velocity: 0.62831853071795862
+    has_acceleration_limits: false
+    max_acceleration: 0
+  ovis_joint_3:
+    has_velocity_limits: true
+    max_velocity: 0.62831853071795862
+    has_acceleration_limits: false
+    max_acceleration: 0
+  ovis_joint_4:
+    has_velocity_limits: true
+    max_velocity: 0.83775804095727813
+    has_acceleration_limits: false
+    max_acceleration: 0
+  ovis_joint_5:
+    has_velocity_limits: true
+    max_velocity: 0.83775804095727813
+    has_acceleration_limits: false
+    max_acceleration: 0
+  ovis_joint_6:
+    has_velocity_limits: true
+    max_velocity: 0.83775804095727813
+    has_acceleration_limits: false
+    max_acceleration: 0
