+++ src/ovis_moveit/config/moveit_controllers.yaml
@@ -1,0 +1,19 @@
+# MoveIt uses this configuration for controller management
+
+moveit_controller_manager: moveit_simple_controller_manager/MoveItSimpleControllerManager
+
+moveit_simple_controller_manager:
+  controller_names:
+    - arm_controller
+
+  arm_controller:
+    type: FollowJointTrajectory
+    action_ns: follow_joint_trajectory
+    default: true
+    joints:
+      - ovis_joint_1
+      - ovis_joint_2
+      - ovis_joint_3
+      - ovis_joint_4
+      - ovis_joint_5
+      - ovis_joint_6
