+++ src/ovis_moveit/config/ovis.srdf
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--This does not replace URDF, and is not an extension of URDF.
+    This is a format for representing semantic information about the robot structure.
+    A URDF file must exist for this robot as well, where the joints and the links that are referenced are defined
+-->
+<robot name="ovis">
+    <!--GROUPS: Representation of a set of joints and links. This can be useful for specifying DOF to plan for, defining arms, end effectors, etc-->
+    <!--LINKS: When a link is specified, the parent joint of that link (if it exists) is automatically included-->
+    <!--JOINTS: When a joint is specified, the child link of that joint (which will always exist) is automatically included-->
+    <!--CHAINS: When a chain is specified, all the links along the chain (including endpoints) are included in the group. Additionally, all the joints that are parents to included links are also included. This means that joints along the chain and the parent joint of the base link are included in the group-->
+    <!--SUBGROUPS: Groups can also be formed by referencing to already defined group names-->
+    <group name="arm">
+        <joint name="virtual_joint"/>
+        <joint name="ovis_joint_base"/>
+        <joint name="ovis_joint_1"/>
+        <joint name="ovis_joint_2"/>
+        <joint name="ovis_joint_3"/>
+        <joint name="ovis_joint_4"/>
+        <joint name="ovis_joint_5"/>
+        <joint name="ovis_joint_6"/>
+        <joint name="ovis_joint_end_effector"/>
+    </group>
+    <!--GROUP STATES: Purpose: Define a named state for a particular group, in terms of joint values. This is useful to define states like 'folded arms'-->
+    <group_state name="ready" group="arm">
+        <joint name="ovis_joint_1" value="0"/>
+        <joint name="ovis_joint_2" value="0.8"/>
+        <joint name="ovis_joint_3" value="-0.5"/>
+        <joint name="ovis_joint_4" value="0"/>
+        <joint name="ovis_joint_5" value="0"/>
+        <joint name="ovis_joint_6" value="0"/>
+    </group_state>
+    <!--VIRTUAL JOINT: Purpose: this element defines a virtual joint between a robot link and an external frame of reference (considered fixed with respect to the robot)-->
+    <virtual_joint name="virtual_joint" type="fixed" parent_frame="world" child_link="base_link"/>
+</robot>
