+++ src/ovis_moveit/config/ovis.urdf.xacro
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<robot xmlns:xacro="http://www.ros.org/wiki/xacro" name="ovis">
+    <xacro:arg name="initial_positions_file" default="initial_positions.yaml" />
+
+    <!-- Import ovis urdf file -->
+    <xacro:include filename="$(find ovis_description)/urdf/ovis_standalone.urdf.xacro" />
+
+    <!-- Import control_xacro -->
+    <xacro:include filename="ovis.ros2_control.xacro" />
+
+
+    <xacro:ovis_ros2_control name="FakeSystem" initial_positions_file="$(arg initial_positions_file)"/>
+
+</robot>
