+++ src/ovis_moveit/config/ros2_controllers.yaml
@@ -1,0 +1,26 @@
+# This config file is used by ros2_control
+controller_manager:
+  ros__parameters:
+    update_rate: 100  # Hz
+
+    arm_controller:
+      type: joint_trajectory_controller/JointTrajectoryController
+
+
+    joint_state_broadcaster:
+      type: joint_state_broadcaster/JointStateBroadcaster
+
+arm_controller:
+  ros__parameters:
+    joints:
+      - ovis_joint_1
+      - ovis_joint_2
+      - ovis_joint_3
+      - ovis_joint_4
+      - ovis_joint_5
+      - ovis_joint_6
+    command_interfaces:
+      - position
+    state_interfaces:
+      - position
+      - velocity
