+++ src/ovis_moveit/launch/demo.launch.py
@@ -1,0 +1,8 @@
+from moveit_configs_utils import MoveItConfigsBuilder
+from moveit_configs_utils.launches import generate_demo_launch
+
+
+def generate_launch_description():
+    moveit_config = MoveItConfigsBuilder("ovis", package_name="ovis_moveit").to_moveit_configs()
+
+    return generate_demo_launch(moveit_config)
